Move `TestContext` to `redis-test`
`TestContext` is not only useful in `redis-rs`' own unit tests, but
probably also to users for their own integration tests. So we make it
publicly available through `redis-test`.

Some parts that rely too much on feature flags, and would require too
many changes in `redis-test` stay in `TestContextExt` in `redis/tests`
for now. If there is demand, we can evaluate on a function-by-function
basis whether bringing them over is worth it.

Fixes #2215
