+++ redis-test/Cargo.toml
@@ -19,9 +19,13 @@ futures = { version = "0.3", optional = true }
 tempfile = "3.27.0"
 socket2 = "0.6"
 rand = "0.10"
+rustls = { version = "0.23", optional = true, default-features = false, features = ["ring"] }
 
 [features]
 aio = ["futures", "redis/aio", "redis/tokio-comp"]
+tls-rustls = ["dep:rustls", "rustls/std", "redis/tls-rustls"]
+tls-rustls-insecure = ["redis/tls-rustls-insecure", "tls-rustls"]
+tokio-rustls-comp = ["redis/tokio-rustls-comp"]
 
 [dev-dependencies]
 tokio = { workspace = true }
