+++ redis-test/src/utils.rs
@@ -1,5 +1,16 @@
+#[cfg(feature = "tls-rustls")]
+use redis::ClientTlsConfig;
+use redis::RedisResult;
+#[cfg(feature = "tls-rustls")]
+use redis::TlsCertificates;
 use socket2::{Domain, Socket, Type};
 use std::ffi::OsStr;
+#[cfg(feature = "tls-rustls")]
+use std::fs::File;
+#[cfg(feature = "tls-rustls")]
+use std::io::BufReader;
+#[cfg(feature = "tls-rustls")]
+use std::io::Read;
 use std::io::Write;
 use std::net::{SocketAddr, TcpListener};
 use std::path::PathBuf;
@@ -403,4 +414,70 @@ pub fn get_random_available_port() -> u16 {
         }
     }
     panic!("Couldn't get a valid port");
+}
+
+pub fn start_tls_crypto_provider() {
+    #[cfg(feature = "tls-rustls")]
+    if rustls::crypto::CryptoProvider::get_default().is_none() {
+        // we don't care about success, because failure means that the provider was set from another thread.
+        let _ = rustls::crypto::ring::default_provider().install_default();
+    }
+}
+
+#[cfg(feature = "tls-rustls")]
+pub fn load_certs_from_file(tls_file_paths: &TlsFilePaths) -> TlsCertificates {
+    let ca_file = File::open(&tls_file_paths.ca_crt).expect("Cannot open CA cert file");
+    let mut root_cert_vec = Vec::new();
+    BufReader::new(ca_file)
+        .read_to_end(&mut root_cert_vec)
+        .expect("Unable to read CA cert file");
+
+    let cert_file = File::open(&tls_file_paths.redis_crt).expect("Cannot open cert file");
+    let mut client_cert_vec = Vec::new();
+    BufReader::new(cert_file)
+        .read_to_end(&mut client_cert_vec)
+        .expect("Unable to read cert file");
+
+    let key_file = File::open(&tls_file_paths.redis_key).expect("Cannot open key file");
+    let mut client_key_vec = Vec::new();
+    BufReader::new(key_file)
+        .read_to_end(&mut client_key_vec)
+        .expect("Unable to read key file");
+
+    TlsCertificates {
+        client_tls: Some(ClientTlsConfig {
+            client_cert: client_cert_vec,
+            client_key: client_key_vec,
+        }),
+        root_cert: Some(root_cert_vec),
+    }
+}
+
+#[cfg(feature = "tls-rustls")]
+pub fn build_single_client<T: redis::IntoConnectionInfo>(
+    connection_info: T,
+    tls_file_params: &Option<TlsFilePaths>,
+    mtls_enabled: bool,
+) -> RedisResult<redis::Client> {
+    if mtls_enabled && tls_file_params.is_some() {
+        redis::Client::build_with_tls(
+            connection_info,
+            load_certs_from_file(
+                tls_file_params
+                    .as_ref()
+                    .expect("Expected certificates when `tls-rustls` feature is enabled"),
+            ),
+        )
+    } else {
+        redis::Client::open(connection_info)
+    }
+}
+
+#[cfg(not(feature = "tls-rustls"))]
+pub fn build_single_client<T: redis::IntoConnectionInfo>(
+    connection_info: T,
+    _tls_file_params: &Option<TlsFilePaths>,
+    _mtls_enabled: bool,
+) -> RedisResult<redis::Client> {
+    redis::Client::open(connection_info)
 }
