+++ redis/tests/support/mod.rs
@@ -2,31 +2,30 @@
 
 #[cfg(feature = "aio")]
 use futures::Future;
+#[allow(
+    unused_imports,
+    reason = "`RedisResult` is used in at least 4 unrelated conditions. So instead of modelling all of them in an unreadable condition, we accept if the import is unused"
+)]
+use redis::RedisResult;
 #[cfg(feature = "cache-aio")]
 use redis::caching::CacheConfig;
 #[cfg(feature = "tls-rustls")]
 use redis::{ClientTlsConfig, TlsCertificates};
-use redis::{
-    Commands, ConnectionAddr, ErrorKind, Pipeline, ProtocolVersion, RedisResult, ServerErrorKind,
-    Value,
-};
+use redis::{Pipeline, Value};
 #[cfg(feature = "aio")]
 use redis::{aio, cmd};
-use redis_test::server::{
-    Module, RedisServer, RedisServerBuilder, RedisServerCommand, use_protocol,
-};
+use redis_test::TestContext;
+#[cfg(feature = "tls-rustls")]
 use redis_test::utils::TlsFilePaths;
-use std::path::PathBuf;
+#[cfg(feature = "tls-rustls")]
+use redis_test::utils::load_certs_from_file;
+
+use std::io;
 #[cfg(feature = "tls-rustls")]
 use std::{
     fs::File,
     io::{BufReader, Read},
 };
-use std::{io, thread::sleep, time::Duration};
-
-#[macro_use]
-mod version;
-pub use version::*;
 
 pub fn current_thread_runtime() -> tokio::runtime::Runtime {
     let mut builder = tokio::runtime::Builder::new_current_thread();
@@ -141,248 +140,42 @@ mod sentinel;
 #[allow(unused_imports)]
 pub use self::sentinel::*;
 
-/// A builder for [`TestContext`]
-///
-/// # Example
-///
-/// ```rust,no_run
-/// use tests::support::TestContextBuilder;
-///
-/// let ctx = TestContextBuilder::new().module(Module::Json).build();
-/// let connection = ctx.connection();
-/// // Use `connection` to run commands
-/// ```
-// Note that this builder is an owned-builder as we want to build in a single chain anyway and do
-// not have to build multiple instances from the same builder. Also, this spares us cloning
-// considerations.
-#[derive(Default)]
-pub struct TestContextBuilder {
-    server_builder: RedisServerBuilder,
-}
-
-impl TestContextBuilder {
-    /// Starts a fresh builder
-    pub fn new() -> Self {
-        Default::default()
-    }
-
-    pub fn address(mut self, address: ConnectionAddr) -> Self {
-        self.server_builder = self.server_builder.address(address);
-        self
-    }
-
-    pub fn config(mut self, config_file: PathBuf) -> Self {
-        self.server_builder = self.server_builder.config(config_file);
-        self
-    }
-
-    pub fn cert_auth_field(mut self, cert_auth_field: impl Into<String>) -> Self {
-        self.server_builder = self.server_builder.cert_auth_field(cert_auth_field);
-        self
-    }
-
-    pub fn cert_auth_field_opt(mut self, opt_cert_auth_field: Option<impl Into<String>>) -> Self {
-        self.server_builder = self.server_builder.cert_auth_field_opt(opt_cert_auth_field);
-        self
-    }
-
-    pub fn module(mut self, module: Module) -> Self {
-        self.server_builder = self.server_builder.module(module);
-        self
-    }
-
-    pub fn modules(mut self, modules: &[Module]) -> Self {
-        self.server_builder = self.server_builder.modules(modules);
-        self
-    }
-
-    pub fn mtls(mut self, enable_mtls: bool) -> Self {
-        self.server_builder = self.server_builder.mtls(enable_mtls);
-        self
-    }
-
-    pub fn tls_paths(mut self, tls_paths: TlsFilePaths) -> Self {
-        self.server_builder = self.server_builder.tls_paths(tls_paths);
-        self
-    }
-
-    pub fn tls_paths_opt(mut self, opt_tls_paths: Option<TlsFilePaths>) -> Self {
-        self.server_builder = self.server_builder.tls_paths_opt(opt_tls_paths);
-        self
-    }
-
-    /// Builds the [`TestContext`] for this instance
-    pub fn build(self) -> TestContext {
-        self.refine_and_build(|_| {})
-    }
-
-    /// Builds the [`TestContext`] for this instance after refining the arguments for the server
-    ///
-    /// # Arguments
-    ///
-    /// * `refiner` - See [`RedisServerBuilder::refine_and_build`]
-    pub fn refine_and_build(self, refiner: impl FnOnce(&mut RedisServerCommand)) -> TestContext {
-        let server = self.server_builder.refine_and_build(refiner);
-        TestContext::from_server(server)
-    }
-}
-
-/// Utility wrapper for a standalone Redis server instance for testing.
-///
-/// # Example
-///
-/// Use `default()` to build a [`TestContext`] with default settings:
-///
-/// ```rust,no_run
-/// use tests::support::TestContext;
-///
-/// let ctx = TestContext::default();
-/// let connection = ctx.connection();
-/// // Use `connection` to run commands
-/// ```
-///
-/// If you need a custom setup, use [`TestContextBuilder`]:
-///
-/// ```rust,no_run
-/// use tests::support::TestContextBuilder;
-///
-/// let ctx = TestContextBuilder::new().module(Module::Json).build();
-/// let connection = ctx.connection();
-/// // Use `connection` to run commands
-/// ```
-pub struct TestContext {
-    pub server: RedisServer,
-    pub client: redis::Client,
-    pub protocol: ProtocolVersion,
-}
-
-pub(crate) fn start_tls_crypto_provider() {
-    #[cfg(feature = "tls-rustls")]
-    if rustls::crypto::CryptoProvider::get_default().is_none() {
-        // we don't care about success, because failure means that the provider was set from another thread.
-        let _ = rustls::crypto::ring::default_provider().install_default();
-    }
-}
-
-impl Default for TestContext {
-    fn default() -> Self {
-        TestContextBuilder::new().build()
-    }
-}
-
-impl TestContext {
-    /// Builds a new instance from a [`RedisServer`]
-    // We intentionally do _not_ implement `From<RedisServer>` as that would be public.
-    //
-    // Instead, users should to go through `TestContextBuilder` to limit the points of entry and
-    // hence help us with maintenance.
-    fn from_server(server: RedisServer) -> Self {
-        let client =
-            build_single_client(server.connection_info(), &server.tls_paths, server.mtls).unwrap();
-
-        if server.tls_paths.is_some() {
-            start_tls_crypto_provider();
-        }
-
-        let mut con;
-
-        let millisecond = Duration::from_millis(1);
-        let mut retries = 0;
-        loop {
-            match client.get_connection() {
-                Err(err) => {
-                    if err.is_connection_refusal() {
-                        sleep(millisecond);
-                        retries += 1;
-                        if retries > 100000 {
-                            panic!(
-                                "Tried to connect too many times, last error: {err}, logfile: {:?}",
-                                server.log_file_contents()
-                            );
-                        }
-                    } else {
-                        panic!(
-                            "Could not connect: {err}, logfile: {:?}",
-                            server.log_file_contents()
-                        );
-                    }
-                }
-                Ok(x) => {
-                    con = x;
-                    break;
-                }
-            }
-        }
-
-        // Redis may still be loading its dataset after accepting connections,
-        // especially with TLS where the handshake completes before Redis is fully ready.
-        // Retry flushdb if the BusyLoading error is returned to allow time for initialization.
-        let mut flush_retries = 0;
-        loop {
-            match con.flushdb::<()>() {
-                Ok(_) => break,
-                Err(err)
-                    if matches!(err.kind(), ErrorKind::Server(ServerErrorKind::BusyLoading)) =>
-                {
-                    sleep(millisecond);
-                    flush_retries += 1;
-                    if flush_retries > 10000 {
-                        panic!(
-                            "Redis is still loading after too many retries, last error: {err}, logfile: {:?}",
-                            server.log_file_contents()
-                        );
-                    }
-                }
-                Err(err) => {
-                    panic!(
-                        "Failed to flush database: {err}, logfile: {:?}",
-                        server.log_file_contents()
-                    );
-                }
-            }
-        }
-
-        TestContext {
-            server,
-            client,
-            protocol: use_protocol(),
-        }
-    }
-
-    pub fn connection(&self) -> redis::Connection {
-        self.client.get_connection().unwrap()
-    }
-
-    #[cfg(feature = "aio")]
-    pub async fn async_connection(&self) -> RedisResult<redis::aio::MultiplexedConnection> {
-        self.client.get_multiplexed_async_connection().await
-    }
-
-    #[cfg(feature = "aio")]
-    pub async fn async_pubsub(&self) -> RedisResult<redis::aio::PubSub> {
-        self.client.get_async_pubsub().await
-    }
-
-    pub fn stop_server(&mut self) {
-        self.server.stop();
-    }
-
+/// Extension of [`TestContext`] taylored to the flags available directly on `redis-rs`.
+pub trait TestContextExt {
     #[cfg(feature = "tokio-comp")]
-    pub async fn multiplexed_async_connection_tokio(
+    async fn multiplexed_async_connection_tokio(
+        &self,
+    ) -> RedisResult<redis::aio::MultiplexedConnection>;
+
+    #[cfg(all(feature = "aio", feature = "cache-aio"))]
+    fn async_connection_with_cache(
+        &self,
+    ) -> impl Future<Output = redis::RedisResult<redis::aio::MultiplexedConnection>>;
+
+    #[cfg(all(feature = "aio", feature = "cache-aio"))]
+    fn async_connection_with_cache_config(
+        &self,
+        cache_config: CacheConfig,
+    ) -> impl Future<Output = redis::RedisResult<redis::aio::MultiplexedConnection>>;
+}
+
+impl TestContextExt for TestContext {
+    #[cfg(feature = "tokio-comp")]
+    async fn multiplexed_async_connection_tokio(
         &self,
     ) -> RedisResult<redis::aio::MultiplexedConnection> {
         self.client.get_multiplexed_async_connection().await
     }
 
     #[cfg(all(feature = "aio", feature = "cache-aio"))]
-    pub fn async_connection_with_cache(
+    fn async_connection_with_cache(
         &self,
     ) -> impl Future<Output = redis::RedisResult<redis::aio::MultiplexedConnection>> {
         self.async_connection_with_cache_config(CacheConfig::default())
     }
 
     #[cfg(all(feature = "aio", feature = "cache-aio"))]
-    pub fn async_connection_with_cache_config(
+    fn async_connection_with_cache_config(
         &self,
         cache_config: CacheConfig,
     ) -> impl Future<Output = redis::RedisResult<redis::aio::MultiplexedConnection>> {
@@ -396,13 +189,6 @@ impl TestContext {
                 )
                 .await
         }
-    }
-}
-
-impl TestContextVersioning for TestContext {
-    fn get_available_components(&self) -> AvailableComponents {
-        let mut conn = self.connection();
-        AvailableComponents::from(&mut conn)
     }
 }
 
@@ -496,55 +282,6 @@ where
 }
 
 #[cfg(feature = "tls-rustls")]
-pub fn load_certs_from_file(tls_file_paths: &TlsFilePaths) -> TlsCertificates {
-    let ca_file = File::open(&tls_file_paths.ca_crt).expect("Cannot open CA cert file");
-    let mut root_cert_vec = Vec::new();
-    BufReader::new(ca_file)
-        .read_to_end(&mut root_cert_vec)
-        .expect("Unable to read CA cert file");
-
-    let cert_file = File::open(&tls_file_paths.redis_crt).expect("Cannot open cert file");
-    let mut client_cert_vec = Vec::new();
-    BufReader::new(cert_file)
-        .read_to_end(&mut client_cert_vec)
-        .expect("Unable to read cert file");
-
-    let key_file = File::open(&tls_file_paths.redis_key).expect("Cannot open key file");
-    let mut client_key_vec = Vec::new();
-    BufReader::new(key_file)
-        .read_to_end(&mut client_key_vec)
-        .expect("Unable to read key file");
-
-    TlsCertificates {
-        client_tls: Some(ClientTlsConfig {
-            client_cert: client_cert_vec,
-            client_key: client_key_vec,
-        }),
-        root_cert: Some(root_cert_vec),
-    }
-}
-
-#[cfg(feature = "tls-rustls")]
-pub(crate) fn build_single_client<T: redis::IntoConnectionInfo>(
-    connection_info: T,
-    tls_file_params: &Option<TlsFilePaths>,
-    mtls_enabled: bool,
-) -> RedisResult<redis::Client> {
-    if mtls_enabled && tls_file_params.is_some() {
-        redis::Client::build_with_tls(
-            connection_info,
-            load_certs_from_file(
-                tls_file_params
-                    .as_ref()
-                    .expect("Expected certificates when `tls-rustls` feature is enabled"),
-            ),
-        )
-    } else {
-        redis::Client::open(connection_info)
-    }
-}
-
-#[cfg(feature = "tls-rustls")]
 pub(crate) fn build_single_client_with_separate_client_cert<T: redis::IntoConnectionInfo>(
     connection_info: T,
     tls_file_params: &TlsFilePaths,
@@ -581,15 +318,6 @@ pub(crate) fn build_single_client_with_separate_client_cert<T: redis::IntoConnec
             root_cert: Some(root_cert_vec),
         },
     )
-}
-
-#[cfg(not(feature = "tls-rustls"))]
-pub(crate) fn build_single_client<T: redis::IntoConnectionInfo>(
-    connection_info: T,
-    _tls_file_params: &Option<TlsFilePaths>,
-    _mtls_enabled: bool,
-) -> RedisResult<redis::Client> {
-    redis::Client::open(connection_info)
 }
 
 #[cfg(feature = "tls-rustls")]
