+++ redis/tests/test_async.rs
@@ -16,6 +16,9 @@ mod basic_async {
         RedisConnectionInfo, RedisError, RedisResult, ScanOptions, ServerErrorKind, Value,
         aio::ConnectionLike, cmd, pipe,
     };
+    use redis_test::TestContext;
+    #[cfg(any(feature = "json", feature = "connection-manager"))]
+    use redis_test::TestContextBuilder;
     use redis_test::redis_value;
     #[cfg(feature = "json")]
     use redis_test::server::Module;
@@ -1649,6 +1652,7 @@ mod basic_async {
     #[cfg(feature = "tls-rustls")]
     mod mtls_test {
         use super::*;
+        use redis_test::utils::build_single_client;
 
         #[rstest::rstest]
         #[cfg_attr(feature = "tokio-comp", case::tokio(RuntimeType::Tokio))]
@@ -1656,10 +1660,7 @@ mod basic_async {
         fn test_should_connect_mtls(#[case] runtime: RuntimeType) {
             let ctx = TestContextBuilder::new().mtls(true).build();
 
-            let client =
-                build_single_client(ctx.server.connection_info(), &ctx.server.tls_paths, true)
-                    .unwrap();
-            let connect = client.get_multiplexed_async_connection();
+            let connect = ctx.client.get_multiplexed_async_connection();
             block_on_all(
                 async move {
                     let mut con = connect.await.unwrap();
