+++ redis/tests/test_cluster_async.rs
@@ -6,6 +6,8 @@ mod support;
 mod cluster_async {
     #[cfg(feature = "tls-rustls")]
     use redis_test::cluster::ClusterType;
+    #[cfg(feature = "tls-rustls")]
+    use redis_test::utils::load_certs_from_file;
     use std::{
         collections::HashMap,
         sync::{
@@ -34,8 +36,13 @@ mod cluster_async {
     };
     use redis::{PushInfo, PushKind, cluster_async::ClusterConnection};
     use redis_test::cluster::{RedisCluster, RedisClusterConfiguration};
-    use redis_test::redis_value;
     use redis_test::server::use_protocol;
+    use redis_test::utils::build_single_client;
+    use redis_test::{
+        REDIS_CE_7_0, TestContextVersioning, VALKEY_9_0, redis_value,
+        run_test_if_version_supported, skip_if_context_does_not_support,
+    };
+
     use test_macros::async_test;
     use tokio::{join, sync::mpsc::UnboundedReceiver};
